requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
accelerate=0.20.3=pypi_0
aiofiles=23.2.1=pypi_0
altair=5.2.0=pypi_0
annotated-types=0.6.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=4.3.0=pypi_0
appdirs=1.4.4=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
attrs=23.2.0=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
bitsandbytes=0.37.0=pypi_0
braceexpand=0.1.7=pypi_0
ca-certificates=2024.2.2=hbcca054_0
certifi=2024.2.2=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
cmake=3.28.3=pypi_0
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.2.0=pypi_0
cudatoolkit=11.8.0=h6a678d5_0
cycler=0.12.1=pypi_0
debugpy=1.6.7=py39h6a678d5_0
decorator=5.1.1=pyhd8ed1ab_0
decord=0.6.0=pypi_0
docker-pycreds=0.4.0=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
exceptiongroup=1.2.0=pypi_0
executing=2.0.1=pyhd8ed1ab_0
fastapi=0.110.0=pypi_0
ffmpy=0.3.2=pypi_0
filelock=3.13.1=pypi_0
fonttools=4.49.0=pypi_0
fsspec=2024.2.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.42=pypi_0
gradio=3.47.1=pypi_0
gradio-client=0.6.0=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.4=pypi_0
httpx=0.27.0=pypi_0
huggingface-hub=0.18.0=pypi_0
idna=3.6=pypi_0
imageio=2.34.0=pypi_0
importlib-resources=6.1.2=pypi_0
iopath=0.1.10=pypi_0
ipykernel=6.29.3=pyhd33586a_0
ipython=8.12.0=pyh41d4057_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.3=pypi_0
joblib=1.3.2=pypi_0
jsonschema=4.21.1=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter_client=7.3.4=pyhd8ed1ab_0
jupyter_core=5.7.2=py39hf3d152e_0
kiwisolver=1.4.5=pypi_0
lazy-loader=0.3=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
lit=17.0.6=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.7.0=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mpmath=1.3.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.2.1=pypi_0
nltk=3.8.1=pypi_0
numpy=1.26.4=pypi_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-cupti-cu11=11.7.101=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-curand-cu11=10.2.10.91=pypi_0
nvidia-cusolver-cu11=11.4.0.1=pypi_0
nvidia-cusparse-cu11=11.7.4.91=pypi_0
nvidia-nccl-cu11=2.14.3=pypi_0
nvidia-nvtx-cu11=11.7.91=pypi_0
omegaconf=2.3.0=pypi_0
opencv-python=4.7.0.72=pypi_0
openssl=3.0.13=h7f8727e_2
orjson=3.9.15=pypi_0
packaging=23.2=pypi_0
pandas=2.2.1=pypi_0
parso=0.8.4=pyhd8ed1ab_0
peft=0.2.0=pypi_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.2.0=pypi_0
pip=23.3.1=py39h06a4308_0
platformdirs=4.2.2=pyhd8ed1ab_0
portalocker=2.8.2=pypi_0
progressbar2=4.4.1=pypi_0
prompt-toolkit=3.0.42=pyha770c72_0
prompt_toolkit=3.0.42=hd8ed1ab_0
protobuf=4.25.3=pypi_0
psutil=5.9.4=pypi_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pydantic=2.6.3=pypi_0
pydantic-core=2.16.3=pypi_0
pydub=0.25.1=pypi_0
pygments=2.18.0=pyhd8ed1ab_0
pyparsing=3.1.1=pypi_0
python=3.9.18=h955ad1f_0
python-dateutil=2.9.0.post0=pypi_0
python-multipart=0.0.9=pypi_0
python-utils=3.8.2=pypi_0
python_abi=3.9=2_cp39
pytz=2024.1=pypi_0
pyyaml=6.0=pypi_0
pyzmq=25.1.2=py39h6a678d5_0
readline=8.2=h5eee18b_0
referencing=0.33.0=pypi_0
regex=2022.10.31=pypi_0
requests=2.31.0=pypi_0
rpds-py=0.18.0=pypi_0
safetensors=0.4.2=pypi_0
scikit-image=0.22.0=pypi_0
scikit-learn=1.4.1.post1=pypi_0
scipy=1.12.0=pypi_0
semantic-version=2.10.0=pypi_0
sentence-transformers=2.2.2=pypi_0
sentencepiece=0.2.0=pypi_0
sentry-sdk=1.40.6=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=68.2.2=py39h06a4308_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.1=pypi_0
sniffio=1.3.1=pypi_0
sqlite=3.41.2=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
starlette=0.36.3=pypi_0
sympy=1.12=pypi_0
threadpoolctl=3.3.0=pypi_0
tifffile=2024.2.12=pypi_0
timm=0.6.13=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.2=pypi_0
toolz=0.12.1=pypi_0
torch=2.0.0=pypi_0
torchaudio=2.0.1=pypi_0
torchvision=0.15.1=pypi_0
tornado=6.1=py39hb9d737c_3
tqdm=4.64.1=pypi_0
traitlets=5.14.3=pyhd8ed1ab_0
transformers=4.30.0=pypi_0
triton=2.0.0=pypi_0
typing-extensions=4.10.0=pypi_0
typing_extensions=4.11.0=pyha770c72_0
tzdata=2024.1=pypi_0
urllib3=2.2.1=pypi_0
uvicorn=0.27.1=pypi_0
visual-genome=1.1.1=pypi_0
wandb=0.16.3=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
webdataset=0.2.48=pypi_0
websockets=11.0.3=pypi_0
wheel=0.41.2=py39h06a4308_0
xz=5.4.6=h5eee18b_0
zeromq=4.3.5=h6a678d5_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h5eee18b_0
